chore: restructure vscode providers into proper folders
